Add a throw exception when receiving a short audio file

+++ src/main/java/com/masonord/harmonyhound/telegram/handlers/MediaHandler.java
@@ -2,6 +2,7 @@ package com.masonord.harmonyhound.telegram.handlers;
 
 import com.google.api.services.drive.model.File;
 import com.google.gson.Gson;
+import com.masonord.harmonyhound.exception.FileTooShortException;
 import com.masonord.harmonyhound.exception.UnsupportedMediaType;
 import com.masonord.harmonyhound.response.rapidapi.Sections;
 import com.masonord.harmonyhound.response.telegram.FilePathResponse;
@@ -11,12 +12,15 @@ import com.masonord.harmonyhound.service.GoogleDriveService;
 import com.masonord.harmonyhound.service.RecognizeMediaService;
 import com.masonord.harmonyhound.util.DownloadUtil;
 import com.masonord.harmonyhound.util.FileSystemUtil;
+import com.masonord.harmonyhound.util.LanguageUtil;
 import com.masonord.harmonyhound.util.MediaUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.telegram.telegrambots.meta.api.methods.BotApiMethod;
 import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
 import org.telegram.telegrambots.meta.api.objects.Message;
+
+import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.net.URI;
 import java.net.URISyntaxException;
@@ -44,11 +48,14 @@ public class MediaHandler {
 
     private final Gson gson;
 
+    private final LanguageUtil languageUtil;
+
     public MediaHandler() {
         this.gson = new Gson();
+        this.languageUtil = new LanguageUtil();
     }
 
-    public BotApiMethod<?> answerMessage(Message message) throws IOException, InterruptedException, URISyntaxException, GeneralSecurityException, UnsupportedMediaType {
+    public BotApiMethod<?> answerMessage(Message message) throws IOException, InterruptedException, URISyntaxException, GeneralSecurityException, UnsupportedMediaType, FileTooShortException {
         String chatId = message.getChatId().toString();
         String in, out;
         FilePathResponse response;
@@ -66,6 +73,13 @@ public class MediaHandler {
         out = "downloaded-media/chat_" + chatId + "/" + response.getResult().getFile_path().split("/")[1].split("\\.")[0] + ".wav";
 
         mediaUtil.convertFileToWav(in, out);
+
+        double fileLength = mediaUtil.getAudioDuration(out);
+
+        if (fileLength < 5.00) {
+            throw new FileTooShortException(languageUtil.getProperty("file.too.short"));
+        }
+
         File fileLink = googleDriveService.uploadFile(out, chatId);
         RecognizedSongResponse recognizedAudio = recognizeMediaService.recognizeAudio(fileLink.getWebViewLink());
         googleDriveService.deleteFile(fileLink.getId());
